+++ Frameworks/CleverPush.xcframework/ios-arm64_x86_64-simulator/CleverPush.framework/Headers/CleverPushInstance.h
@@ -17,6 +17,7 @@
 #import "CPSubscription.h"
 #import "CPChannelTag.h"
 #import "CPChannelTopic.h"
+#import "CPIabTcfMode.h"
 
 @interface CPNotificationReceivedResult : NSObject
 
@@ -56,6 +57,7 @@ typedef void (^CPFailureBlock)(NSError* error);
 
 typedef void (^CPAppBannerActionBlock)(CPAppBannerAction* action);
 typedef void (^CPAppBannerShownBlock)(CPAppBanner* appBanner);
+typedef void (^CPAppBannerDisplayBlock)(UIViewController *viewController);
 
 typedef void (^CPLogListener)(NSString* message);
 
@@ -84,6 +86,8 @@ extern NSString * const CLEVERPUSH_SDK_VERSION;
 
 - (void)setTrackingConsentRequired:(BOOL)required;
 - (void)setTrackingConsent:(BOOL)consent;
+- (void)setSubscribeConsentRequired:(BOOL)required;
+- (void)setSubscribeConsent:(BOOL)consent;
 - (void)enableDevelopmentMode;
 - (void)subscribe;
 - (void)subscribe:(CPHandleSubscribedBlock)subscribedBlock;
@@ -102,6 +106,8 @@ extern NSString * const CLEVERPUSH_SDK_VERSION;
 - (void)handleNotificationOpened:(NSDictionary*)messageDict isActive:(BOOL)isActive actionIdentifier:(NSString*)actionIdentifier;
 - (void)handleNotificationReceived:(NSDictionary*)messageDict isActive:(BOOL)isActive;
 - (void)enqueueRequest:(NSURLRequest*)request onSuccess:(CPResultSuccessBlock)successBlock onFailure:(CPFailureBlock)failureBlock;
+- (void)enqueueRequest:(NSURLRequest*)request onSuccess:(CPResultSuccessBlock)successBlock onFailure:(CPFailureBlock)failureBlock withRetry:(BOOL)retryOnFailure;
+- (void)enqueueFailedRequest:(NSURLRequest *)request withRetryCount:(NSInteger)retryCount onSuccess:(CPResultSuccessBlock)successBlock onFailure:(CPFailureBlock)failureBlock;
 - (void)handleJSONNSURLResponse:(NSURLResponse*) response data:(NSData*) data error:(NSError*) error onSuccess:(CPResultSuccessBlock)successBlock onFailure:(CPFailureBlock)failureBlock;
 - (void)addSubscriptionTopic:(NSString*)topicId;
 - (void)addSubscriptionTopic:(NSString*)topicId callback:(void(^)(NSString *))callback;
@@ -165,7 +171,9 @@ extern NSString * const CLEVERPUSH_SDK_VERSION;
 - (void)getAppBannersByGroup:(NSString*)groupId callback:(void(^)(NSMutableArray <CPAppBanner*>*))callback;
 - (void)setAppBannerOpenedCallback:(CPAppBannerActionBlock)callback;
 - (void)setAppBannerShownCallback:(CPAppBannerShownBlock)callback;
+- (void)setShowAppBannerCallback:(CPAppBannerDisplayBlock)callback;
 - (void)setApiEndpoint:(NSString*)apiEndpoint;
+- (void)setIabTcfMode:(CPIabTcfMode)mode;
 - (void)setAuthorizerToken:(NSString*)authorizerToken;
 - (void)setCustomTopViewController:(UIViewController*)viewController;
 - (void)updateBadge:(UNMutableNotificationContent*)replacementContent API_AVAILABLE(ios(10.0));
@@ -190,6 +198,7 @@ extern NSString * const CLEVERPUSH_SDK_VERSION;
 - (NSString*)getApiEndpoint;
 - (NSString*)channelId;
 - (UIViewController*)getCustomTopViewController;
+- (CPIabTcfMode)getIabTcfMode;
 
 - (UIColor*)getBrandingColor;
 
@@ -239,6 +248,7 @@ extern NSString * const CLEVERPUSH_SDK_VERSION;
 - (CPHandleSubscribedBlock)getSubscribeHandler;
 - (void)setSubscribeHandler:(CPHandleSubscribedBlock)subscribedCallback;
 - (void)initFeatures;
+- (void)initIabTcf;
 - (void)initAppReview;
 - (BOOL)hasNewTopicAfterOneHour:(NSDictionary*)config initialDifference:(NSInteger)initialDifference displayDialogDifference:(NSInteger)displayAfter;
 - (NSInteger)secondsAfterLastCheck;
@@ -254,6 +264,10 @@ extern NSString * const CLEVERPUSH_SDK_VERSION;
 - (BOOL)getHasTrackingConsent;
 - (BOOL)getHasTrackingConsentCalled;
 - (void)waitForTrackingConsent:(void(^)(void))callback;
+- (BOOL)getSubscribeConsentRequired;
+- (BOOL)getHasSubscribeConsent;
+- (BOOL)getHasSubscribeConsentCalled;
+- (void)waitForSubscribeConsent:(void(^)(void))callback;
 - (void)addSubscriptionTagToApi:(NSString*)tagId callback:(void (^)(NSString *))callback onFailure:(CPFailureBlock)failureBlock;
 - (void)removeSubscriptionTagFromApi:(NSString*)tagId callback:(void (^)(NSString *))callback onFailure:(CPFailureBlock)failureBlock;
 - (void)initTopicsDialogData:(NSDictionary*)config syncToBackend:(BOOL)syncToBackend;